refactor: combine components
